requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
mysqlclient
python-dotenv
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]==1.7.4
alembic
pydantic[email]
pytz
psutil
ipython
httpx
slowapi